DERBY-2094 (partial) Add the order status business transaction implementation for oe.direct.Standard. This implements the transaction using client side JBDC. Now that populating the database is supported run some test orders status and Stock level transactions in the OETest. Need more validation on the results of the transactions.
git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@499618 13f79535-47bb-0310-9956-ffa450edef68

+++ java/testing/org/apache/derbyTesting/system/oe/model/Order.java
@@ -51,7 +51,7 @@ public class Order {
     private short warehouse;
     private int customer;
     private Timestamp entry_d;
-    private Short carrier_id;
+    private Integer carrier_id; // JDBC maps SMALLINT to java.lang.Integer
     private int ol_cnt;
     private boolean all_local;
 
@@ -61,10 +61,10 @@ public class Order {
     public void setAll_local(boolean all_local) {
         this.all_local = all_local;
     }
-    public Short getCarrier_id() {
+    public Integer getCarrier_id() {
         return carrier_id;
     }
-    public void setCarrier_id(Short carrier_id) {
+    public void setCarrier_id(Integer carrier_id) {
         this.carrier_id = carrier_id;
     }
     public int getCustomer() {
